pyproject.toml
[tool.poetry]
name = "review-ai"
version = "0.1.0"
description = ""
authors = ["AswanthManoj <aswanthmanoj51@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.47.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
jinja2 = "^3.1.4"
python-multipart = "^0.0.12"
httpx = "^0.27.2"
openai = "^1.51.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
